Refactor require AMD loader to Webpack loader with bundles in app/assets/occsn/sticky_booking
