--- assets/scss/booking.scss
@@ -1,171 +1,0 @@
-// contains the CSS brandon used to overwrite the existing styles
-
-$medium-up: 'only screen and (min-width: 992px)';
-
-body {
-  background-color: #191919;
-}
-
-.background-image {
-  opacity: 1;
-  filter: blur(4px);
-
-  &:after {
-    content: '';
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    top: 0;
-    left: 0;
-    background-color: rgba(#191919,0.8);
-  }
-}
-
-.address.vcard {
-  background-image: none !important;
-  padding-left: 0 !important;
-  line-height: 24px !important;
-}
-.merchant-information,
-.product-description {
-  padding-left: 0 !important;
-}
-
-.container-fluid {
-  padding: 0;
-}
-
-// occasion UI changes
-
-
-.btn-primary {
-  background-color: #3C9FF3 !important;
-
-  &:hover {
-    background-color: darken(#3C9FF3,20%) !important;
-  }
-}
-
-
-
-
-
-
-.crazy {
-  background-color: #fff;
-
-
-  @media #{$medium-up} {
-    overflow: hidden;
-  }
-
-  .occasion-wrapper {
-    .occasion-accent-image, .occasion-content {
-      @media #{$medium-up} {
-    		height: 100vh;
-    	}
-
-    }
-  }
-
-  .occasion-accent-image {
-    background-size: cover;
-    background-position: center center;
-    height: 400px;
-
-    @media #{$medium-up} {
-      height: 100vh;
-    }
-  }
-  .occasion-content {
-    overflow-x: scroll;
-
-    .layout-wrapper {
-      max-width: 700px;
-      width: 100%;
-      margin: 0 auto;
-      padding: 40px;
-    }
-  }
-
-
-  .product-title {
-    margin-bottom: 20px;
-    font-size: 40px;
-    line-height: 50px;
-  }
-  .product-desc {
-    border-bottom: 1px dotted #ccc;
-    padding-bottom: 30px;
-    margin-bottom: 40px !important;
-
-    p {
-      font-size: 16px;
-      line-height: 28px;
-    }
-
-    .product-location {
-      font-weight: bold;
-      margin-top: 40px;
-      text-align: center;
-
-      @media #{$medium-up} {
-        text-align: left;
-      }
-
-      svg {
-        width: 18px;
-        height: auto;
-        position: block;
-        margin: 0 auto;
-
-        @media #{$medium-up} {
-          display: inline-block;
-          position: relative;
-          top: 3px;
-        }
-      }
-
-      span {
-        display: block;
-
-        @media #{$medium-up} {
-          display: inline;
-        }
-      }
-    }
-  }
-
-  #occsn_widget {
-    box-shadow: 0 0 0;
-    max-width: none;
-
-    .widget_row {
-      margin-left: 0;
-      border-left: 1px solid rgba(#ccc,0.5);
-      padding-right: 0;
-    }
-
-    .guider-middle, .guider-top, .guider-bottom {
-      background-image: none;
-      padding: 0 0 0 20px;
-      border-left: 4px solid rgba(#ccc,0.9);
-    }
-
-    h3 {
-      font-weight: bold;
-      font-size: 20px;
-    }
-
-    .btn {
-      font-size: 12px;
-    }
-    .form-split {
-      margin: 10px 0;
-
-      @media #{$medium-up} {
-        margin: 0;
-      }
-    }
-  }
-}
