+++ index.html
@@ -6,9 +6,10 @@
     
     <title>Loading...</title>
 
-    <link rel="stylesheet" media="all" href="node_modules/bootstrap/dist/css/bootstrap.min.css">
-    <link rel="stylesheet" media="all" href="assets/css/booking.css">
-    <link rel="stylesheet" media="all" href="assets/css/animate.css">
+    <script type="application/javascript" src="app/assets/occsn/stickybooking/vendor-bundle.js" ></script>
+    <script type="application/javascript" src="app/assets/occsn/stickybooking/app-bundle.js"></script>
+    <link rel="stylesheet" href="app/assets/occsn/stickybooking/vendor-bundle.css" />
+    <link rel="stylesheet" href="app/assets/occsn/stickybooking/app-bundle.css" />
   </head>
   <body class="booking-ui-body-padding body-scroll-lock">
 
@@ -17,8 +18,8 @@
     <!-- Set Static API Key and Product ID -->
     <script>
       window.OCCSN = {
-        api_key: '463436b5ebf74b42873aaa544801f91a',
-        product_id: 'ahtdt9cutawrxuwieabgyq'
+        api_key: '59a8517bcbfa3fd9458b46d4ee59f8fd717fb80d',
+        product_id: 'okf16c2huihqtypfmml80q'
       };
       console.log("window.OCCSN", window.OCCSN);
     </script>
@@ -26,33 +27,10 @@
     <!-- Font-Awesome -->
     <script src="https://use.fontawesome.com/c385cd600f.js"></script>
 
-    <!-- Node Dependencies -->
-    <script src="require/require.js"></script>
-    <script src="require/require-config.js"></script>
-
-    <!-- Date Helper Functions -->
-    <script src="assets/js/date-helpers.js"></script>
-
-    
     <!-- Spreedly iFrame API -->
     <script src="https://core.spreedly.com/iframe/iframe-v1.min.js"></script>
 
     <!-- SqPaymentForm library -->
     <script type="text/javascript" src="https://js.squareup.com/v2/paymentform"></script>
-
-
-    <!-- Angular Files -->
-    <!-- App Module -->
-    <script src="/app/app.module.js"></script>
-
-    <!-- Occasion SDK Service -->
-    <script src="/app/services/occasionSDK.service.js"></script>
-
-    <!-- Booking Component -->
-    <script src="/app/components/booking/booking.component.js"></script>
-
-    <!-- Calendar Component -->
-    <script src="/app/components/calendar/calendar.component.js"></script>
-
   </body>
 </html>
