+++ package.json
@@ -18,8 +18,17 @@
     "jquery": "^3.2.1",
     "occasion-sdk": "git+https://github.com/nicklandgrebe/occasion-sdk-js#ccb1381"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "css-loader": "^0.28.7",
+    "expose-loader": "^0.7.3",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "html-loader": "^0.5.1",
+    "ngtemplate-loader": "^2.0.1",
+    "style-loader": "^0.19.0",
+    "webpack": "^3.8.1"
+  },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack --config webpack.config.js"
   }
 }
