+++ src/components/booking/booking.component.html
@@ -1,7 +1,7 @@
 <main class="booking-page-container">
 
     <div class="loadingAnimation" ng-if="displayLoading">
-        <img src="assets/images/loading_icon.gif" />
+        <img src="" />
     </div>
 
     <div class="booking-sidebar">
