+++ src/libs/booking.css
@@ -289,11 +289,6 @@ th {
   .table-bordered td {
     border: 1px solid #ddd !important; } }
 
-@font-face {
-  font-family: 'Glyphicons Halflings';
-  src: url("../fonts/bootstrap/glyphicons-halflings-regular.eot");
-  src: url("../fonts/bootstrap/glyphicons-halflings-regular.eot?#iefix") format("embedded-opentype"), url("../fonts/bootstrap/glyphicons-halflings-regular.woff2") format("woff2"), url("../fonts/bootstrap/glyphicons-halflings-regular.woff") format("woff"), url("../fonts/bootstrap/glyphicons-halflings-regular.ttf") format("truetype"), url("../fonts/bootstrap/glyphicons-halflings-regular.svg#glyphicons_halflingsregular") format("svg"); }
-
 .glyphicon {
   position: relative;
   top: 1px;
@@ -6044,7 +6039,6 @@ This file is generated by `grunt build`, do not edit it by hand.
   display: block;
   width: 12px;
   height: 12px;
-  background: url("chosen-sprite.png") -42px 1px no-repeat;
   font-size: 1px; }
 
 .chosen-container-single .chosen-single abbr:hover {
@@ -6064,8 +6058,7 @@ This file is generated by `grunt build`, do not edit it by hand.
 .chosen-container-single .chosen-single div b {
   display: block;
   width: 100%;
-  height: 100%;
-  background: url("chosen-sprite.png") no-repeat 0px 2px; }
+  height: 100%; }
 
 .chosen-container-single .chosen-search {
   position: relative;
@@ -6081,8 +6074,6 @@ This file is generated by `grunt build`, do not edit it by hand.
   height: auto;
   outline: 0;
   border: 1px solid #aaa;
-  background: white url("chosen-sprite.png") no-repeat 100% -20px;
-  background: url("chosen-sprite.png") no-repeat 100% -20px;
   font-size: 1em;
   font-family: sans-serif;
   line-height: normal;
@@ -6345,8 +6336,6 @@ This file is generated by `grunt build`, do not edit it by hand.
 
 .chosen-rtl .chosen-search input[type="text"] {
   padding: 4px 5px 4px 20px;
-  background: white url("chosen-sprite.png") no-repeat -30px -20px;
-  background: url("chosen-sprite.png") no-repeat -30px -20px;
   direction: rtl; }
 
 .chosen-rtl.chosen-container-single .chosen-single div b {
@@ -6365,7 +6354,6 @@ This file is generated by `grunt build`, do not edit it by hand.
   .chosen-container-multi .chosen-choices .search-choice .search-choice-close,
   .chosen-container .chosen-results-scroll-down span,
   .chosen-container .chosen-results-scroll-up span {
-    background-image: url("chosen-sprite@2x.png") !important;
     background-size: 52px 37px !important;
     background-repeat: no-repeat !important; } }
